+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [\#75](https://github.com/Finschia/wasmd/pull/75) test: add the test case for InactiveContract
 * [\#74](https://github.com/Finschia/wasmd/pull/74) add event checking to TestInstantiateContract2
 * [\#78](https://github.com/Finschia/wasmd/pull/78) add the check for TestMigrateContract
+* [\#69](https://github.com/Finschia/wasmd/pull/69) refactor: refactor test cases for Params
 
 ### Bug Fixes
 * [\#62](https://github.com/Finschia/wasmd/pull/62) fill ContractHistory querier result's Updated field